test(ai): split growth_stage_planner_test under per-file line budget (#3371)
The growth-stage planner test file exceeded the repo-lint 1000 non-comment
line limit (1064). Extract shared fixtures into a non-test support file and
move the Builder relocation / anti-thrash (AC7, AC14) groups into a dedicated
test file. No test behavior change.

Refs #3371
